Initial setup of the AMARL project
The dependencies and conda environments have been setup to be able to
start working on the project. The external
sequential_social_dilemma_games requires conda, hence this is now a
conda project as well. Still I use Makefiles for automating the setup,
as external projects depend on other external projects and the setup is
not that trivial. This way it is easy to replicate and documented.