Implement rollout worker
This will be needed by the A2C algorithm to rollout trajectories of an
agent within a specified environment. Encapsulating this in a class
makes it more modular and readable. Handling multiple environments has
been handed over to a special environment wrapper following the pattern
of the OpenAI baseline. This better separates concerns and makes it
easier to extend the rollout later on for multi-agent environments.