Improved test coverage of gym wrappers
I forgot to test the episodic return wrapper. Also removed the flaky
marker from the policies tests, because the current version of rerun
failures causes problems with the latest pytest. I've decided to remove
the library entirely and work with a fixed seed instead. This test
should still fail in case the algorithm is not learning anymore no
matter the seed used.