Skip renderer tests on devices without a display
The travis build would fail those tests because of the missing X
session. To fix this I decided to simply skip the tests on travis, as
the wrapper is not that important anymore. In the future I might remove
it completely. There is also the option of using a virtual render buffer
on travis, but for only 2 tests this seems to be overkill.