Add dirty first working LSTM policy
The LSTM policy that has been implemented works conceptually, but is in
dire need of a cleanup. Also it might be good to approach the issue
again from a test driven point of view. Currently having more than one
worker agent is not supported either, servery limiting the use of the
policy.