Env done message and stop on done worker config
The rollout worker class has been updated to allow cancelling the
rollout when an environment has terminated. This behaviour is necessary
when wanting to train policies making use of LSTMs as they need to train
on the final steps to propagate errors back in time properly. Also a
environment has terminated message has been added, so LSTM policies can
subscribe to it and reset their hidden state when an environment has
terminated.