Implement multi env reset properly
Keeping track and asserting on terminated environments within the multi
environment wrapper, also requires to reactivate the terminated
environments after a reset, as well as providing a proper interface to
do so. This fixes some bugs introduced by the new concept that have been
highlighted by the worker tests failing.