Update rollout workers to allow for selective envs
The next step in supporting irregular rollout necessary for LSTM based
actor-critic models, was to let the rollout workers handle terminated
environments properly and to record data on a per rank basis. This way
it is possible to pass rollouts of different lengths to the policy. This
required new support for rank ids in the rollout logic.