Refactor policy tests and worker pipeline
To make its easier to extend the LSTM actor critic algorithm with
multiple environment support the policy tests have been refactored. This
should make them more expressive and easier to understand. While working
on this refactoring I also found some opportunities simplifying the
rollout handling. By gathering all the data and graph nodes used by
training in one place and having the actual actions used by the
environment its own entity as a numpy array it removed some of the logic
necessary to move data from and to the GPU in case it was used.