Implement video visualisation
With this it is possible to send video snippets from episodes to
tensorboard which helps evaluating the agents. Also changed the policy
interface to expect torch tensor, which prevents casting issues between
training on multiple environments and rolling out an inference
trajectory on just one environment for visualisation. Should also
prevent back and forth casting and make the interface cleaner in
general.